Aplicación terminada!

+++ uaiContacts/src/main/java/uaiContacts/repository/PlaceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import uaiContacts.model.Place;
 									 
-public interface PlaceRepository extends CrudRepository<Place, Integer>{
+public interface PlaceRepository extends CrudRepository<Place, Long>{
 	public List<Place> findByPhoneLike(String phone);
 
 }
